+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [REFACTORING] The `DynamicButtonStyle` is now an enum
 - [ADD] Each style must now adopt the `DynamicButtonBuildable` protocol
 - [ADD] The `DynamicButtonPathVector` struct
+- [ADD] `styleWithName` method in the `DynamicButtonStyle`
 
 ## [Version 3.1.0](https://github.com/yannickl/DynamicButton/releases/tag/3.1.0)
 Released on 2016-11-05.